app/requirements.txt
fastapi==0.99.1
uvicorn==0.23.2
pydantic==1.10.12
chromadb==0.4.4
numpy==1.24.3
more-itertools==10.1.0
webencodings==0.5.1
jinja2==3.1.2
markupsafe==2.1.3
psutil==5.9.5
filelock==3.12.2
lit==16.0.6
omegaconf==2.3.0
hydra-core==1.3.2
langchain==0.0.251
openai==0.27.8
sentence_transformers==2.2.2
huggingface_hub==0.16.4
unstructured==0.10.5
ctransformers[cuda]==0.2.26
pypdf==3.15.4
